+++ src/main/java/com/codetaylor/mc/athenaeum/util/FileHelper.java
@@ -1,0 +1,24 @@
+package com.codetaylor.mc.athenaeum.util;
+
+import java.io.Closeable;
+import java.io.IOException;
+
+public class FileHelper {
+
+  public static void closeSilently(Closeable closeable) {
+
+    if (closeable != null) {
+
+      try {
+        closeable.close();
+      } catch (IOException e) {
+        //
+      }
+    }
+  }
+
+  private FileHelper() {
+
+  }
+
+}